Refactored DialogEditorContextMenu
